+++ crates/uselesskey-core-rustls-pki/Cargo.toml
@@ -1,22 +1,21 @@
 [package]
-name = "uselesskey-rustls"
+name = "uselesskey-core-rustls-pki"
 version = "0.3.0"
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "rustls-pki-types and rustls config adapters for uselesskey X.509/key fixtures."
+description = "rustls-pki-types adapters for uselesskey fixture crates."
 categories.workspace = true
-keywords = ["rustls", "tls", "fixtures", "testing", "adapter"]
+keywords = ["rustls", "pki", "fixtures", "testing", "adapter"]
 readme = "README.md"
 exclude = ["fuzz/**", "corpus/**", "**/*.der", "**/*.pem"]
 homepage.workspace = true
-documentation = "https://docs.rs/uselesskey-rustls"
+documentation = "https://docs.rs/uselesskey-core-rustls-pki"
 authors.workspace = true
 
 [dependencies]
-uselesskey-core-rustls-pki = { path = "../uselesskey-core-rustls-pki", version = "0.3.0", default-features = false }
-rustls = { version = "0.23", default-features = false, features = ["std", "tls12"], optional = true }
+rustls-pki-types = "1"
 
 # Key type crates (all optional)
 uselesskey-x509 = { path = "../uselesskey-x509", version = "0.3.0", optional = true }
@@ -24,30 +23,20 @@ uselesskey-rsa = { path = "../uselesskey-rsa", version = "0.3.0", optional = tru
 uselesskey-ecdsa = { path = "../uselesskey-ecdsa", version = "0.3.0", optional = true }
 uselesskey-ed25519 = { path = "../uselesskey-ed25519", version = "0.3.0", optional = true }
 
-[package.metadata.docs.rs]
-features = ["all", "tls-config", "rustls-ring"]
-
 [features]
 default = ["x509"]
-x509 = ["uselesskey-core-rustls-pki/x509", "dep:uselesskey-x509"]
-rsa = ["uselesskey-core-rustls-pki/rsa", "dep:uselesskey-rsa"]
-ecdsa = ["uselesskey-core-rustls-pki/ecdsa", "dep:uselesskey-ecdsa"]
-ed25519 = ["uselesskey-core-rustls-pki/ed25519", "dep:uselesskey-ed25519"]
+x509 = ["dep:uselesskey-x509"]
+rsa = ["dep:uselesskey-rsa"]
+ecdsa = ["dep:uselesskey-ecdsa"]
+ed25519 = ["dep:uselesskey-ed25519"]
 all = ["x509", "rsa", "ecdsa", "ed25519"]
-server-config = ["dep:rustls", "x509"]
-client-config = ["dep:rustls", "x509"]
-tls-config = ["server-config", "client-config"]
-rustls-ring = ["rustls?/ring"]
-rustls-aws-lc-rs = ["rustls?/aws_lc_rs"]
 
 [dev-dependencies]
-rustls-pki-types = "1"
 uselesskey-core = { path = "../uselesskey-core", version = "0.3.0" }
 uselesskey-x509 = { path = "../uselesskey-x509", version = "0.3.0" }
 uselesskey-rsa = { path = "../uselesskey-rsa", version = "0.3.0" }
 uselesskey-ecdsa = { path = "../uselesskey-ecdsa", version = "0.3.0" }
 uselesskey-ed25519 = { path = "../uselesskey-ed25519", version = "0.3.0" }
-rustls = { version = "0.23", features = ["ring"] }
-hex = "0.4"
-serde.workspace = true
-insta.workspace = true
+
+[package.metadata.docs.rs]
+features = ["all"]
